Am o versiune stabila

+++ delete_product.php
@@ -7,7 +7,8 @@
 		//Sterg si imaginea asociata acelui produs 
 		$result = mysql_query("select image from products where id_product = " . $_GET["product_id"]);
 		while ($row = mysql_fetch_assoc($result))
-			unlink ($row['image']);
+			if ($row['image'] != "img/noImage.jpg")
+				unlink ($row['image']);
 		
 		mysql_query("delete from products where id_product = " . $_GET["product_id"]);
 		header ("Location: index.php?delete=successful");
